1. Radar almost fully working and with it's size and range fully adjustable 2. Resolved some type conversion warnings

+++ Application/Source/SceneShop.cpp
@@ -619,8 +619,8 @@ void SceneShop::Render() {
 
 
 	rotShip.y += 45 * (float)_dt;
-	int spaceShipX = Application::_windowWidth / 20;
-	int spaceShipY = Application::_windowHeight / 15;
+	float spaceShipX = (float)Application::_windowWidth / 20;
+	float spaceShipY = (float)Application::_windowHeight / 15;
 
 	Vector3 scaleShip(35, 35, 35);
